tests/wpt/web-platform-tests/css/cssom-view/offsetTopLeft-empty-inline-offset.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/cssom-view-1/#extensions-to-the-htmlelement-interface">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
  .container {
    position: relative;
    width: 8em;
    height: 7em;
    padding: 1em;
  }
</style>
<div class="container" style="writing-mode:horizontal-tb;">
  offset
  <br>offset<span class="target"></span><span>ref</span>
</div>
<div class="container" style="writing-mode:vertical-lr;">
  offset
  <br>offset<span class="target"></span><span>ref</span>
</div>
<div class="container" style="writing-mode:vertical-rl;">
  offset
  <br>offset<span class="target"></span><span>ref</span>
</div>
<div class="container" style="writing-mode:horizontal-tb; direction:rtl;">
  offset
  <br>offset<span class="target"></span><span>ref</span>
</div>
<div class="container" style="writing-mode:vertical-lr; direction:rtl;">
  offset
  <br>offset<span class="target"></span><span>ref</span>
</div>
<div class="container" style="writing-mode:vertical-rl; direction:rtl;">
  offset
  <br>offset<span class="target"></span><span>ref</span>
</div>
<script>
var i = 0;
document.querySelectorAll('span.target').forEach((target) => {
  var ref = target.nextSibling;
  test(() => {
    assert_equals(target.offsetLeft, ref.offsetLeft, 'offsetLeft');
    assert_equals(target.offsetTop, ref.offsetTop, 'offsetTop');
  }, 'offsetTop/Left of empty inline elements should work as if they were not empty: ' + i);
  i++;
});
</script>
